Create rowsreader/split.go
